Mofidied namespace to ConnectionConfig
