--- src/lib/y2network/connection/ethernet.rb
@@ -1,9 +1,0 @@
-require "y2network/connection/connection"
-
-module Y2Network
-  module Connection
-    # Profile settings for ethernet connections
-    class Ethernet < Config
-    end
-  end
-end
