+++ src/lib/y2network/connection_config/ethernet.rb
@@ -1,0 +1,9 @@
+require "y2network/connection_config/base"
+
+module Y2Network
+  module ConnectionConfig
+    # Configuration for ethernet connections
+    class Ethernet < Base
+    end
+  end
+end
