+++ src/lib/y2network/connection_config/wireless.rb
@@ -1,9 +1,28 @@
-require "y2network/connection/connection"
+# Copyright (c) [2019] SUSE LLC
+#
+# All Rights Reserved.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms of version 2 of the GNU General Public License as published
+# by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, contact SUSE LLC.
+#
+# To contact SUSE LLC about this file by physical or electronic mail, you may
+# find current contact information at www.suse.com.
+
+require "y2network/connection_config/base"
 
 module Y2Network
-  module Connection
-    # Profile settings for wireless connections
-    class Wireless < Config
+  module ConnectionConfig
+    # Configuration for wireless connections
+    class Wireless < Base
       # wireless options
       # @return [String]
       attr_accessor :mode
